Khuyến mãi (OLP K.Chuyên 2012)

+++ UPCODER/SP.cpp
@@ -2,12 +2,12 @@
 using namespace std;
 
 int main() {
-    int m, n, t;
+    int64_t m, n, t;
     cin >> m >> n >> t;
 
-    int x = n / (m + 1);
-    int y = n % (m + 1);
-    int ans = (x * t * m) + (y * t);
+    int64_t x = n / (m + 1);
+    int64_t y = n % (m + 1);
+    int64_t ans = (x * t * m) + (y * t);
 
     cout << ans;
     return 0;
